+++ docs/test/chunk-2RFPKQPS.js
@@ -14,7 +14,7 @@ import {
 } from "./chunk-MV6BA3M2.js";
 import {
   AuthService
-} from "./chunk-B23OTVUL.js";
+} from "./chunk-LKVJC54M.js";
 import {
   NavigationService
 } from "./chunk-UA7UPQID.js";
@@ -521,4 +521,4 @@ var ProviderPageComponent = class _ProviderPageComponent {
 export {
   ProviderPageComponent
 };
-//# sourceMappingURL=chunk-6O5V75C7.js.map
+//# sourceMappingURL=chunk-2RFPKQPS.js.map